Move updated inelastic operator to its own class
