+++ LoKI-B/Operators.h
@@ -138,7 +138,6 @@ namespace loki {
          *                 only electronic, or also vibrational, rotational?
          */
         void evaluateInelasticOperators(const Grid& grid, const EedfMixture& mixture);
-        void evaluateInelasticOperatorsNew(const Grid& grid, const Vector& eedf, const EedfMixture& mixture);
         Matrix inelasticMatrix;
     };
